% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_accurate_dprime.R
\name{compute_accurate_dprime}
\alias{compute_accurate_dprime}
\title{Run the matlab code to use the accurate d' code.}
\usage{
compute_accurate_dprime(model_import, scale_factor = 1)
}
\arguments{
\item{scale_factor}{}
}
\value{

}
\description{
Run the matlab code to use the accurate d' code.
}
