% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map_parameters.R
\name{create_map_parameters}
\alias{create_map_parameters}
\title{Create the parameters used to define the efficiency maps.}
\usage{
create_map_parameters()
}
\value{

}
\description{
Create the parameters used to define the efficiency maps.
}
