% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_retinal_constants.R
\name{create_retinal_constants}
\alias{create_retinal_constants}
\title{create_retinal_constants The function used to generate the retinal parameters that determine the number of GCs on a ring.}
\usage{
create_retinal_constants(version = "standard")
}
\arguments{
\item{scale.dist}{}
}
\value{

}
\description{
create_retinal_constants The function used to generate the retinal parameters that determine the number of GCs on a ring.
}
