% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_optimal_criterion.R
\name{custom_add_threshold}
\alias{custom_add_threshold}
\title{Determine the threshold cutoff to use for determining false hits.}
\usage{
custom_add_threshold(x, spacing, patch_rad)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Determine the threshold cutoff to use for determining false hits.
}
