% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_efficiency_maps.R
\name{estimate_efficiency_maps}
\alias{estimate_efficiency_maps}
\title{Main function to estimate the efficiency maps used for search}
\usage{
estimate_efficiency_maps(
  map_parameters,
  optim_params = list(n_cores = 1, maxit = 3000, tolerance = 1e-12, NP = NULL)
)
}
\arguments{
\item{optim_params}{}
}
\value{

}
\description{
Main function to estimate the efficiency maps used for search
}
