% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_alternate_criterion.R
\name{find_alternate_criterion}
\alias{find_alternate_criterion}
\title{Find alternate criterion}
\usage{
find_alternate_criterion(model_search)
}
\arguments{
\item{model_search}{}
}
\value{

}
\description{
Find alternate criterion
}
