% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_fit_criterion.R
\name{find_best_fit_criterion}
\alias{find_best_fit_criterion}
\title{Find the criterion that best correlates the model observer with human search.}
\usage{
find_best_fit_criterion(model, human)
}
\arguments{
\item{human}{}
}
\value{

}
\description{
Find the criterion that best correlates the model observer with human search.
}
