% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_optimal_criterion.R
\name{find_optimal_criterion}
\alias{find_optimal_criterion}
\title{Find optimal criterion for a the max rule searcher}
\usage{
find_optimal_criterion(model_search)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Find optimal criterion for a the max rule searcher
}
