% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_maps.R
\name{fit_maps}
\alias{fit_maps}
\title{fit_maps interpolates a 1d d' efficiency map and produces a 2d map that is used for search.}
\usage{
fit_maps(maps_1d)
}
\arguments{
\item{maps_1d}{}
}
\value{

}
\description{
fit_maps interpolates a 1d d' efficiency map and produces a 2d map that is used for search.
}
