% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_maps_legacy_spacing.R
\name{fit_maps_legacy}
\alias{fit_maps_legacy}
\title{fit_maps interpolates a 1d d' efficiency map and produces a 2d map that is used for search.}
\usage{
fit_maps_legacy(maps_1d)
}
\arguments{
\item{spacing_scalar}{}
}
\value{

}
\description{
fit_maps interpolates a 1d d' efficiency map and produces a 2d map that is used for search.
}
