% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_single_search.R
\name{format_single_search}
\alias{format_single_search}
\title{Take a data frame of raw search responses and format the data for fitting and presentation.}
\usage{
format_single_search(raw_search)
}
\arguments{
\item{raw_search}{}
}
\value{

}
\description{
Take a data frame of raw search responses and format the data for fitting and presentation.
}
