% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_fit_map.R
\name{get_single_gain_map}
\alias{get_single_gain_map}
\title{Create a single map and save it.}
\usage{
get_single_gain_map(
  a,
  b_2,
  c,
  efficiency,
  prior_type,
  params_detection,
  neural_resource
)
}
\arguments{
\item{neural_resource}{}
}
\value{

}
\description{
Create a single map and save it.
}
