% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_maps.R, R/fit_maps_legacy_spacing.R
\name{interpolate_single_map}
\alias{interpolate_single_map}
\title{Single interpolated map}
\usage{
interpolate_single_map(spacing, scalar, spacing_full)

interpolate_single_map(spacing, scalar, spacing_full)
}
\arguments{
\item{spacing_full}{}
}
\value{



}
\description{
Single interpolated map

Single interpolated map
}
