% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_figures.R
\name{plot_full_summary}
\alias{plot_full_summary}
\title{Full summary}
\usage{
plot_full_summary(response_df, parameter_label = "Intercept")
}
\arguments{
\item{parameter_label}{}
}
\value{

}
\description{
Full summary
}
