% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_single_search.R
\name{run_single_search}
\alias{run_single_search}
\title{Run a single iteration of the search algorithm, in MATLAB}
\usage{
run_single_search(
  map_path,
  efficiency,
  contrast,
  n_trials,
  radius,
  priorType,
  search_params,
  seed_val = 1,
  single_thread = FALSE
)
}
\arguments{
\item{search_params}{}
}
\value{

}
\description{
Run a single iteration of the search algorithm, in MATLAB
}
