% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchr_clean_pii.R
\name{schr_clean_pii}
\alias{schr_clean_pii}
\title{Standardizes names, emails, and phone numbers}
\usage{
schr_clean_pii(x, type = "name")
}
\arguments{
\item{x}{a vector containing email, phone data, customer or business names}

\item{type}{input data type of c('email', 'phone', 'name')}
}
\value{
a vector containing standardized email, phone data, customer or business names
}
\description{
Performs simple preparation of email, customer name, and phone data by standardizing
input data for US locale and removing common problems specific to those data types.
Run prior to calculating combinatorial matching
}
\examples{
schr_clean_pii(c("Sun100@yahoo.com","RICHY5000@MSN.COM","ab?c@mac.com"), 'email')
schr_clean_pii('12345678910', 'phone')
}
