% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchr_extract_extension.R
\name{schr_extract_extension}
\alias{schr_extract_extension}
\title{Gets email or phone extension}
\usage{
schr_extract_extension(emails, ext_separator)
}
\arguments{
\item{emails}{vector of emails or phone numbers}

\item{ext_separator}{regex separator characters denoting email or phone number extension e.g. ext_separator = "ext", "#", "@"}
}
\value{
email or phone number extensions as char vec
}
\description{
Returns extension from email or phone number when given a separation character like "@"
}
\examples{
schr_extract_extension(c("Sun100@yahoo.com","RICHY5000@MSN.COM","ab?c@mac.com"), '@')
}
