% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchr_remove_extension.R
\name{schr_remove_extension}
\alias{schr_remove_extension}
\title{Remove email or phone extension}
\usage{
schr_remove_extension(emails, ext_separator)
}
\arguments{
\item{emails}{vector of emails or phone numbers}

\item{ext_separator}{regex separator characters denoting email or phone number extension e.g. ext_separator = "ext", ext_separator = '#', '@'}
}
\value{
email or phone number extensions as char vec
}
\description{
Removes extension from email or phone number when given a separation character like "@"
}
\examples{
schr_remove_extension(c("Sun100@yahoo.com","RICHY5000@MSN.COM","ab?c@mac.com"), '@')
}
