% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchr_table_characters.R
\name{schr_tbl_chars}
\alias{schr_tbl_chars}
\title{Tables character counts}
\usage{
schr_tbl_chars(char_vec)
}
\arguments{
\item{char_vec}{vector of character strings to table with character counts}
}
\value{
tabled and aligned character count matrix
}
\description{
Counts and tables characters in a vector of strings ordered from most to least
frequent
}
\examples{
schr_tbl_chars(c(" Amazon.com Inc ", " amazon inc"))
}
