% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchr_validate_emails.R
\name{schr_validate_emails}
\alias{schr_validate_emails}
\title{Check if an email looks mis-typed}
\usage{
schr_validate_emails(email_vec)
}
\arguments{
\item{email_vec}{character vector of email addresses}
}
\value{
a bool vector where T means okay, and FALSE means probably a mis-typed email
}
\description{
Checks for common email issues, but email conformity is hard. See
tradeoffs discussion for regexs:  https://www.regular-expressions.info/email.html
}
\examples{
schr_validate_emails(c("Sun100@yahoo.com","RICHY5000@MSN.COM","ab?c@mac.com"))
}
