% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_best_fit_search.R
\name{select_best_fit_search}
\alias{select_best_fit_search}
\title{Returns a data frame with a subset of conditions select that are the best fit selections.}
\usage{
select_best_fit_search(human_model_df)
}
\arguments{
\item{human_model_df}{}
}
\value{

}
\description{
Returns a data frame with a subset of conditions select that are the best fit selections.
}
