% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_xy.R
\name{sp_xy}
\alias{sp_xy}
\title{sp_xy Ganglion cell spacing function.}
\usage{
sp_xy(x, y, spacing.scalar = 1)
}
\arguments{
\item{spacing.scalar}{}
}
\value{

}
\description{
This function returns the spacing between retinal ganglion cell receptive fields. The x,y
coordinates are specified in degrees of visual angle. The reference point 0,0 is the center
of the visual field; the center of the fovea centralis.
}
