% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_maps_disk.R
\name{store_maps_disk}
\alias{store_maps_disk}
\title{Function to store fitted maps on disk.}
\usage{
store_maps_disk(file_id, maps_2d_lst)
}
\arguments{
\item{file_id}{}
}
\value{

}
\description{
Function to store fitted maps on disk.
}
