% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_search.R
\name{subset_search}
\alias{subset_search}
\title{A wrapper for selecting a subset of relevant search.}
\usage{
subset_search(search_df)
}
\arguments{
\item{search_df}{}
}
\value{

}
\description{
A wrapper for selecting a subset of relevant search.
}
