% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_search.R
\name{summary_search}
\alias{summary_search}
\title{Formats the search data into a marginaled response table. Summary search statistics.}
\usage{
summary_search(imported_search)
}
\arguments{
\item{imported_search}{}
}
\value{

}
\description{
Formats the search data into a marginaled response table. Summary search statistics.
}
