% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_search_spacing.R
\name{summary_search_spacing}
\alias{summary_search_spacing}
\title{Formats the search data into a marginaled response table. Summary search statistics. Use spacing rather than eccentricity.}
\usage{
summary_search_spacing(imported_search, bins = NA, nbins = 5)
}
\arguments{
\item{imported_search}{}
}
\value{

}
\description{
Formats the search data into a marginaled response table. Summary search statistics. Use spacing rather than eccentricity.
}
