% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_2d_map.R
\name{visualize_2d_map}
\alias{visualize_2d_map}
\title{List of 2d maps, output a list of figure objects.}
\usage{
visualize_2d_map(map_2d)
}
\arguments{
\item{map_2d}{}
}
\value{

}
\description{
List of 2d maps, output a list of figure objects.
}
