% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_human.R
\name{visualize_human}
\alias{visualize_human}
\title{Plot a figure summarizing human results}
\usage{
visualize_human(human_raw)
}
\arguments{
\item{human_summary_data}{}
}
\value{

}
\description{
Plot a figure summarizing human results
}
