% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson_spacing.R
\name{watson_spacing}
\alias{watson_spacing}
\title{Watson spacing function from Watson (2014) JoV.}
\usage{
watson_spacing(x, y, binoc = 1, sp_scal = 1, f0 = 0)
}
\arguments{
\item{y}{}
}
\value{

}
\description{
Watson spacing function from Watson (2014) JoV.
}
