% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_maps.R
\name{interpolate_single_map}
\alias{interpolate_single_map}
\title{Single interpolated map}
\usage{
interpolate_single_map(eccentricity, scalar, eccentricity_full)
}
\arguments{
\item{scalar}{}
}
\value{

}
\description{
Single interpolated map
}
